refactor(codex): consolidate catalog test fixtures
