+++ extensions/codex/src/session-catalog.test-helpers.ts
@@ -1,4 +1,11 @@
-// Shared fixtures for the split Codex session catalog suites.
+// Register narrow mocks before any production imports evaluate the catalog graph.
+// oxfmt-ignore
+import {
+  commandRpcMocks,
+  pinnedConnectionMocks,
+  transcriptMirrorMocks,
+  nodeHostMocks,
+} from "./session-catalog.test-mocks.js";
 import { createHash } from "node:crypto";
 import fsSync from "node:fs";
 import fs from "node:fs/promises";
@@ -19,7 +26,7 @@ import type { PluginRuntime } from "openclaw/plugin-sdk/plugin-runtime";
 import type { SessionCatalogProvider as RegisteredSessionCatalogProvider } from "openclaw/plugin-sdk/session-catalog";
 import { resolveStorePath } from "openclaw/plugin-sdk/session-store-runtime";
 import { withEnvAsync } from "openclaw/plugin-sdk/test-env";
-import { vi } from "vitest";
+import { afterEach, beforeEach, vi } from "vitest";
 import {
   resolveCodexAppServerHomeDir,
   resolveCodexAppServerLocalHomeDir,
@@ -61,8 +68,28 @@ export const CODEX_NODE_CONTINUE_COMMANDS = [
   CODEX_APP_SERVER_THREAD_TURNS_LIST_COMMAND,
   CODEX_CLI_SESSION_RESUME_COMMAND,
 ] as const;
-export const originalPath = process.env.PATH;
+const originalPath = process.env.PATH;
 export const tempDirs: string[] = [];
+
+beforeEach(() => {
+  nodeHostMocks.runNodePtyCommand.mockClear();
+  nodeHostMocks.userShellPaths.clear();
+  commandRpcMocks.codexControlRequest.mockReset();
+  pinnedConnectionMocks.getClient.mockReset();
+  pinnedConnectionMocks.getClient.mockResolvedValue(pinnedConnectionMocks.client);
+  pinnedConnectionMocks.releaseClient.mockReset();
+  pinnedConnectionMocks.request.mockReset();
+  transcriptMirrorMocks.importCodexThreadHistoryToTranscript.mockReset();
+  transcriptMirrorMocks.importCodexThreadHistoryToTranscript.mockResolvedValue({
+    importedMessages: 0,
+    omittedMessages: 0,
+  });
+});
+
+afterEach(async () => {
+  process.env.PATH = originalPath;
+  await Promise.all(tempDirs.splice(0).map((dir) => fs.rm(dir, { recursive: true, force: true })));
+});
 
 const archiveLocalCodexSession = codexSessionCatalogRuntime.archiveLocal;
 const continueLocalCodexSessionRuntime = codexSessionCatalogRuntime.continueLocal;
@@ -544,6 +571,10 @@ export function createGatewayApi(runtime: PluginRuntime, apiConfig: OpenClawConf
 }
 
 export {
+  commandRpcMocks,
+  pinnedConnectionMocks,
+  transcriptMirrorMocks,
+  nodeHostMocks,
   fs,
   fsSync,
   os,
